+++ modules/compute/emr-serverless/app.py
@@ -12,7 +12,7 @@ deployment_name = os.getenv("SEEDFARMER_DEPLOYMENT_NAME")
 module_name = os.getenv("SEEDFARMER_MODULE_NAME")
 
 if len(f"{project_name}-{deployment_name}") > 36:
-    raise Exception("This module cannot support a project+deployment name character length greater than 35")
+    raise ValueError("This module cannot support a project+deployment name character length greater than 35")
 
 
 def _param(name: str) -> str:
@@ -24,10 +24,10 @@ vpc_id = os.getenv(_param("VPC_ID"))  # required
 private_subnet_ids = json.loads(os.getenv(_param("PRIVATE_SUBNET_IDS"), ""))  # required
 
 if not vpc_id:
-    raise Exception("Missing input parameter vpc-id")
+    raise ValueError("Missing input parameter vpc-id")
 
 if not private_subnet_ids:
-    raise Exception("Missing input parameter private-subnet-ids")
+    raise ValueError("Missing input parameter private-subnet-ids")
 
 app = App()