+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### **Added**
 
 - enforced TLS version 1.2, node-node encryption and encryption at rest on OS module
- 
+- added `emr-serverless` module with unit-tests
+- added workflow entries to all IDF modules
+
 ### **Changed**
 
 ### **Removed**