+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - replaced exporting metadata with seedfarmer command
 - storage/buckets - added `usedforsecurity=False` to the sha1 creation of bucket names
+- applying changes based off security scans and code standards recommendations
 
 ### **Removed**