+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+### **Changed**
+
+### **Removed**
+
+=======
+
+## v1.2.0 (2023-11-09)
+
+### **Added**
+
 - added `sagemaker-studio` module with unit-tests
 - enforced TLS version 1.2, node-node encryption and encryption at rest on OS module
 - added `emr-serverless` module with unit-tests